+++ src/main/java/com/handwoong/rainbowletter/domain/letter/controller/LetterController.java
@@ -1,16 +1,14 @@
-package com.handwoong.rainbowletter.controller.letter;
+package com.handwoong.rainbowletter.domain.letter.controller;
 
+import com.handwoong.rainbowletter.domain.letter.dto.ReplyRequestDto;
+import com.handwoong.rainbowletter.domain.letter.service.LetterService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import com.handwoong.rainbowletter.dto.letter.ReplyRequestDto;
-import com.handwoong.rainbowletter.service.letter.LetterService;
-
-import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequiredArgsConstructor
