+++ src/main/java/com/handwoong/rainbowletter/domain/letter/dto/AirtableUpdateDto.java
@@ -1,4 +1,4 @@
-package com.handwoong.rainbowletter.dto.letter;
+package com.handwoong.rainbowletter.domain.letter.dto;
 
 public record AirtableUpdateDto(
         String GPT,
