+++ src/main/java/com/handwoong/rainbowletter/domain/letter/dto/AirtableUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.handwoong.rainbowletter.dto.letter;
+package com.handwoong.rainbowletter.domain.letter.dto;
 
 public record AirtableUpdateRequestDto(
         AirtableUpdateDto fields
