+++ src/main/java/com/handwoong/rainbowletter/domain/letter/dto/ChatGptResponseDto.java
@@ -1,4 +1,4 @@
-package com.handwoong.rainbowletter.dto.letter;
+package com.handwoong.rainbowletter.domain.letter.dto;
 
 import java.util.List;
