+++ src/main/java/com/handwoong/rainbowletter/domain/letter/dto/ReplyRequestDto.java
@@ -1,4 +1,4 @@
-package com.handwoong.rainbowletter.dto.letter;
+package com.handwoong.rainbowletter.domain.letter.dto;
 
 public record ReplyRequestDto(
         String baseId,
