+++ src/main/java/com/handwoong/rainbowletter/domain/letter/service/LetterService.java
@@ -1,0 +1,7 @@
+package com.handwoong.rainbowletter.domain.letter.service;
+
+import com.handwoong.rainbowletter.domain.letter.dto.ReplyRequestDto;
+
+public interface LetterService {
+    void reply(final ReplyRequestDto replyRequest);
+}
