--- src/main/java/com/handwoong/rainbowletter/service/letter/LetterService.java
@@ -1,7 +1,0 @@
-package com.handwoong.rainbowletter.service.letter;
-
-import com.handwoong.rainbowletter.dto.letter.ReplyRequestDto;
-
-public interface LetterService {
-    void reply(final ReplyRequestDto replyRequest);
-}
